Create README.md
Hi, I'm stevecripe57
I was sent here to solve a question and issue I'm having with my Firefox Browser for Android.
I would appreciate any help with the issue of how I transfer my Collections on my old phone Firefox browser to my new phone Firefox browser. Firefox apparently has been working on this issue since 2020, and that is the answer they're still giving to the many people on their forum with the same problem. So it's obviously not a priority for them. Any suggestions?